Replaced int by float in puissance.c

+++ TP2/src/puissance.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 
-int power(int a, int b){
-  int result = 1;
+float power(int a, int b){
+  float result = 1;
 
   for(int i = 0; i < (b < 0 ? -b : b); i++){
     result *= a;
@@ -11,13 +11,14 @@ int power(int a, int b){
 }
 
 int main(){
-  printf("Test 2 & 3 : %d\n", power(2, 3));
-  printf("Test 2 & -3 : %d\n", power(2, -3));
-  printf("Test 2 & 0 : %d\n", power(2, 0));
-  printf("Test -2 & 3 : %d\n", power(-2, 3));
-  printf("Test -2 & -3 : %d\n", power(-2, -3));
+  printf("Test 2 & 3 : %f\n", power(2, 3));
+  printf("Test 2 & -3 : %f\n", power(2, -3));
+  printf("Test 2 & 0 : %f\n", power(2, 0));
+  printf("Test -2 & 3 : %f\n", power(-2, 3));
+  printf("Test -2 & -3 : %f\n", power(-2, -3));
   
   return 0;
 }
+
